+++ components/bsx/Offer/MyOffer.vue
@@ -97,7 +97,7 @@ import OfferMixin from '~/utils/mixins/offerMixin'
 import { tokenIdToRoute } from '@/components/unique/utils'
 
 const components = {
-  Identity: () => import('@/components/shared/format/Identity.vue'),
+  Identity: () => import('@/components/shared/identity/Identity.vue'),
   Money: () => import('@/components/shared/format/Money.vue'),
 }
