+++ components/bsx/Offer/OffersUserTable.vue
@@ -96,7 +96,7 @@ import { notificationTypes, showNotification } from '~/utils/notification'
 import { tokenIdToRoute } from '~/components/unique/utils'
 
 const components = {
-  Identity: () => import('@/components/shared/format/Identity.vue'),
+  Identity: () => import('@/components/shared/identity/Identity.vue'),
   Money: () => import('@/components/shared/format/Money.vue'),
   BasicSwitch: () => import('@/components/shared/form/BasicSwitch.vue'),
 }
