+++ components/rmrk/Profile/ProfileLink.vue
@@ -39,7 +39,7 @@ import PrefixMixin from '@/utils/mixins/prefixMixin'
 import shortAddress from '@/utils/shortAddress'
 
 const components = {
-  Identity: () => import('@/components/shared/format/Identity.vue'),
+  Identity: () => import('@/components/shared/identity/Identity.vue'),
   LinkResolver: () => import('@/components/shared/LinkResolver.vue'),
 }
