+++ components/rmrk/Profile/Sales.vue
@@ -97,7 +97,7 @@ import { Event } from '../service/types'
 import PrefixMixin from '~/utils/mixins/prefixMixin'
 
 const components = {
-  Identity: () => import('@/components/shared/format/Identity.vue'),
+  Identity: () => import('@/components/shared/identity/Identity.vue'),
   Pagination: () => import('@/components/rmrk/Gallery/Pagination.vue'),
   BlockExplorerLink: () => import('@/components/shared/BlockExplorerLink.vue'),
 }
