+++ components/sales/SalesTable.vue
@@ -105,7 +105,7 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 import PrefixMixin from '~/utils/mixins/prefixMixin'
 
 const components = {
-  Identity: () => import('@/components/shared/format/Identity.vue'),
+  Identity: () => import('@/components/shared/identity/Identity.vue'),
   Money: () => import('@/components/shared/format/Money.vue'),
   Loader: () => import('@/components/shared/Loader.vue'),
   BasicImage: () => import('@/components/shared/view/BasicImage.vue'),
