+++ components/spotlight/SpotlightTable.vue
@@ -213,7 +213,7 @@ import PrefixMixin from '~/utils/mixins/prefixMixin'
 type Address = string | GenericAccountId | undefined
 
 const components = {
-  Identity: () => import('@/components/shared/format/Identity.vue'),
+  Identity: () => import('@/components/shared/identity/Identity.vue'),
   Money: () => import('@/components/shared/format/Money.vue'),
   SpotlightDetail: () => import('./SpotlightDetail.vue'),
   Loader: () => import('@/components/shared/Loader.vue'),
